fix 3

+++ src/encode/blp/ctx/write_blp_full_jpegs.rs
@@ -97,7 +97,7 @@ impl EncoderCtx {
                         .as_ref()
                         .map(|s| if m.included { s.scan_len } else { 0 })
                 })
-                .map(|v| v as usize)
+                .map(|v| v)
                 .sum()
         } else {
             self.mips
